cmake: Added ZI and INCREMENTAL flags to allow for edit & continue on windows and MSVC. Got rid of unused BETATEST cmake configuration which was never used. str: renamed STR macro to STR_LIT which encodes string literal as a str_t base64: changed to size_t as the size type instead of int. types: removed md_vec3_soa, this structure was not really needed and apis are better off without this. util: Unified com operations under two public procedures spatial_hash: resolved issue where a bounding box which is exactly divisible by 6.0Å caused issues. This will be changed in the future.

+++ src/core/md_arena_allocator.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 struct md_allocator_i;
@@ -8,16 +9,16 @@ struct md_allocator_i;
 
 typedef struct md_vm_arena_t {
     void* base;     // Base adress of reserved memory space
-    uint64_t size;  // Reservation size
-    uint64_t pos;
-    uint64_t commit_pos;
-    uint64_t align;
+    size_t size;  // Reservation size
+    size_t pos;
+    size_t commit_pos;
+    size_t align;
     uint64_t magic;
 } md_vm_arena_t;
 
 typedef struct md_vm_arena_temp_t {
     md_vm_arena_t* arena;
-    uint64_t pos;
+    size_t pos;
 } md_vm_arena_temp_t;
 
 #ifdef __cplusplus
@@ -27,7 +28,7 @@ extern "C" {
 // This is a simple allocator which does not free any memory until reset or destroy is called, and then all memory is freed at once.
 // Internally it allocates memory in pages from a backing allocator to keep up with the demand. All pages are of course freed upon reset or destroy.
 
-struct md_allocator_i* md_arena_allocator_create(struct md_allocator_i* backing, uint64_t page_size);
+struct md_allocator_i* md_arena_allocator_create(struct md_allocator_i* backing, size_t page_size);
 void md_arena_allocator_reset(struct md_allocator_i* arena);
 void md_arena_allocator_destroy(struct md_allocator_i* arena);
 
@@ -35,18 +36,18 @@ void md_arena_allocator_destroy(struct md_allocator_i* arena);
 // Specialized arena allocator which uses resorts to the os virtual memory allocator
 // Should be prefered whenever possible over the very generic version above
 
-void md_vm_arena_init(md_vm_arena_t* vm_arena, uint64_t reservation_size);
+void md_vm_arena_init(md_vm_arena_t* vm_arena, size_t reservation_size);
 void md_vm_arena_free(md_vm_arena_t* vm_arena);
 
-void* md_vm_arena_push(md_vm_arena_t* vm_arena, uint64_t size);
-void* md_vm_arena_push_zero(md_vm_arena_t* vm_arena, uint64_t size);
-void* md_vm_arena_push_aligned(md_vm_arena_t* vm_arena, uint64_t size, uint64_t align);
-void  md_vm_arena_pop(md_vm_arena_t* vm_arena, uint64_t size);
+void* md_vm_arena_push(md_vm_arena_t* vm_arena, size_t size);
+void* md_vm_arena_push_zero(md_vm_arena_t* vm_arena, size_t size);
+void* md_vm_arena_push_aligned(md_vm_arena_t* vm_arena, size_t size, size_t align);
+void  md_vm_arena_pop(md_vm_arena_t* vm_arena, size_t size);
 
 void  md_vm_arena_reset(md_vm_arena_t* vm_arena);
 
-void	 md_vm_arena_set_pos(md_vm_arena_t* vm_arena, uint64_t pos);
-uint64_t md_vm_arena_get_pos(md_vm_arena_t* vm_arena);
+void	 md_vm_arena_set_pos(md_vm_arena_t* vm_arena, size_t pos);
+size_t md_vm_arena_get_pos(md_vm_arena_t* vm_arena);
 
 struct md_allocator_i md_vm_arena_create_interface(md_vm_arena_t* arena);
